+++ src/main/java/roidrole/thaumicinfo/top/ThaumicInformationTOPPlugin.java
@@ -7,8 +7,8 @@ import roidrole.thaumicinfo.ThaumicInformationConfig;
 
 import java.util.function.Function;
 
-//TODO: Arcane Note, Arcane Ear show notes (+ HWYLA), Congrega Mystica
-//TODO: Fix discrepancy in WAILA showing 0 suction and TOP not.
+//TODO: HWYLA: Arcane Ear, Vis Generator, Void syphon, Note Blocks
+//TODO: Merge HWYLA and TOP configs
 public class ThaumicInformationTOPPlugin {
 	public static int SUCTION_ELEMENT_ID = -1;
 	public static int CONTENT_ELEMENT_ID = -1;
@@ -20,29 +20,39 @@ public class ThaumicInformationTOPPlugin {
 	public static class Initializer implements Function<ITheOneProbe, Void> {
 		@Override
 		public Void apply(ITheOneProbe probe) {
-			SUCTION_ELEMENT_ID = probe.registerElementFactory(new ElementSuction.Factory());
-			CONTENT_ELEMENT_ID = probe.registerElementFactory(new ElementContent.Factory());
-
 			if(ThaumicInformationConfig.topConfig.essentiaTransport){
+				SUCTION_ELEMENT_ID = probe.registerElementFactory(new ElementSuction.Factory());
+				CONTENT_ELEMENT_ID = probe.registerElementFactory(new ElementContent.Factory());
+
 				probe.registerProvider(ProviderEssentiaTransport.INSTANCE);
 			}
 
+			if(ThaumicInformationConfig.topConfig.arcaneEar){
+				probe.registerProvider(ProviderArcaneEar.INSTANCE);
+			}
+
 			if(ThaumicInformationConfig.topConfig.gogglesDisplay){
-				/*
-				registrar.registerBodyProvider(ProviderGogglesDisplay.INSTANCE, IGogglesDisplayExtended.class);
-				*/
+				probe.registerProvider(ProviderGogglesDisplay.INSTANCE);
 			}
 
 			if(ThaumicInformationConfig.topConfig.visBattery) {
-				/*
-				registrar.registerBodyProvider(ProviderBlockVisBattery.INSTANCE, BlockVisBattery.class);
-				*/
+				probe.registerProvider(ProviderVisBattery.INSTANCE);
+			}
+
+			if(ThaumicInformationConfig.topConfig.visGenerator){
+				probe.registerProvider(ProviderVisGenerator.INSTANCE);
+			}
+
+			if(ThaumicInformationConfig.topConfig.voidSiphon){
+				probe.registerProvider(ProviderVoidSiphon.INSTANCE);
 			}
 
 			if(ThaumicInformationConfig.topConfig.brainInJar) {
-				/*
-				registrar.registerBodyProvider(ProviderBrainJar.INSTANCE, TileJarBrain.class);
-				*/
+				probe.registerProvider(ProviderBrainJar.INSTANCE);
+			}
+
+			if(ThaumicInformationConfig.topConfig.noteBlock){
+				probe.registerProvider(ProviderNote.INSTANCE);
 			}
 			return null;
 		}